Implemented ability to save book preferences and some cleanup on existing apis.
