+++ API/Entities/Interfaces/IBookService.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using API.Parser;
+using HtmlAgilityPack;
 using VersOne.Epub;
 
 namespace API.Entities.Interfaces
